+++ components/forms/Input/react.js
@@ -1,69 +1,50 @@
-import React, { Component } from 'react'
+import React, { memo } from 'react'
 import classNames from 'classnames/bind'
 import { mX } from '@masonite/svg-icons'
 import PropTypes from 'prop-types'
-import s from './style.module.scss'
+import style from './style.module.scss'
 
-const cx = classNames.bind(s)
+const cx = classNames.bind(style)
 
-export default class Input extends Component {
-  constructor(props) {
-    super(props)
+const Input = ({
+  error, onChange, onKeyDown, placeholder, value, variant,
+}) => (
+  <div className={cx('input-wrapper')}>
+    <input
+      className={cx('input', {
+        'input--light': variant === 'light',
+        'input--dark': variant === 'dark',
+        'input--error': error,
+      })}
+      type="text"
+      placeholder={placeholder}
+      onKeyDown={onKeyDown}
+      onChange={ev => onChange(ev.target.value)}
+      value={value}
+    />
+    <div
+      className={cx(['input-icon', { 'input-icon--error': error }])}
+      dangerouslySetInnerHTML={{ __html: mX }}
+    />
+  </div>
+)
 
-    this.state = {
-      value: '',
-    }
-  }
-
-  _onChange = ev => {
-    const { onChange } = this.props
-    this.setState({ value: ev.target.value })
-    onChange(ev.target.value)
-  }
-
-  render() {
-    const {
-      error, onKeyDown, placeholder, variant,
-    } = this.props
-    const { value } = this.state
-
-    return (
-      <div className={cx('input-wrapper')}>
-        <input
-          className={cx('input', {
-            'input--dark': variant === 'dark',
-            'input--error': error,
-          })}
-          type="text"
-          placeholder={placeholder}
-          onKeyDown={onKeyDown}
-          onChange={this._onChange}
-          value={value}
-        />
-        <div
-          className={cx(['input-icon', { 'input-icon--error': error }])}
-          dangerouslySetInnerHTML={{ __html: mX }}
-        />
-      </div>
-    )
-  }
-}
+export default memo(Input)
 
 Input.propTypes = {
-  error: PropTypes.bool,
-  placeholder: PropTypes.string,
-  variant: PropTypes.string,
   onChange: PropTypes.func,
+  error: PropTypes.bool,
   onKeyDown: PropTypes.func,
+  placeholder: PropTypes.string,
+  value: PropTypes.string,
+  variant: PropTypes.oneOf(['dark', 'light']),
 }
 
 Input.defaultProps = {
   onChange: null,
-  onKeyDown: null,
-}
-
-Input.defaultProps = {
-  variant: 'light',
-  placeholder: '',
   error: false,
+  onKeyDown: null,
+  placeholder: '',
+  value: '',
+  variant: 'light',
 }