+++ packages/external-ui-react/CHANGELOG.md
@@ -5,6 +5,42 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.7] - 2019-02-19
+
+### Added
+
+- Dropdown icon choices for FilterableSearch
+
+## [1.5.6] - 2019-02-19
+
+### Added
+
+- Add `buttonLabel` prop
+
+### Changed
+
+- FilterableSearch method names: `onSearch` is now `onSubmit`
+- FilterableSearch's button only appears when there is an `onSubmit` method provided
+
+## [1.5.5] - 2019-02-19
+
+### Fixed
+
+- Build issue
+
+## [1.5.4] - 2019-02-19
+
+### Fixed
+
+- Old build published
+
+## [1.5.3] - 2019-02-18
+
+### Changed
+
+- forms/FilterableSearch to be a PureComponent
+- forms/Input to be a function component
+
 ## [1.5.0] - 2019-02-15
 
 ### Added